Phase out TestDoubles::AddOneAction
