+++ spec/acceptance/orchestrator/execute_spec.rb
@@ -13,10 +13,10 @@ describe LightService::Orchestrator do
 
     def self.steps
       [
-        TestDoubles::AddOneAction,
+        TestDoubles::AddsOneAction,
         execute(->(ctx) { ctx.number += 1 }),
         execute(->(ctx) { ctx[:something] = 'hello' }),
-        TestDoubles::AddOneAction
+        TestDoubles::AddsOneAction
       ]
     end
   end
