+++ spec/acceptance/orchestrator/iterate_spec.rb
@@ -10,14 +10,14 @@ describe LightService::Orchestrator do
     def self.run(context)
       with(context).reduce([
                              iterate(:numbers, [
-                                       TestDoubles::AddOneAction
+                                       TestDoubles::AddsOneAction
                                      ])
                            ])
     end
 
     def self.run_single(context)
       with(context).reduce([
-                             iterate(:numbers, TestDoubles::AddOneAction)
+                             iterate(:numbers, TestDoubles::AddsOneAction)
                            ])
     end
   end
